refactor(assetManagement): 폴더 구조 개편 및 API 리팩토링
useGetAssetStocks 및 관련 API를 widgets 폴더에서 entities로 이동하여 모듈화된 폴더 구조를 개선했습니다.
또한, API 호출 경로를 통합하고 사용성을 높이기 위해 rewrites 설정과 getServiceUrl 로직을 업데이트했습니다.
