+++ frontend/next.config.mjs
@@ -7,6 +7,14 @@ const __dirname = path.dirname(__filename);
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  async rewrites() {
+    return [
+      {
+        source: "/api/:path*", // 프록시를 사용할 경로
+        destination: "http://api.gaemischool.com:8000/api/:path*", // 실제 API 서버 주소
+      },
+    ];
+  },
   webpack: (config) => {
     config.resolve.alias = {
       ...config.resolve.alias,
