+++ frontend/src/entities/assetManagement/apis/getAssetsStock.ts
