+++ frontend/src/shared/utils/getServiceUrl.ts
@@ -1,3 +1,8 @@
+import { SERVICE_SERVER_URL } from "@/shared";
+
 export const getServiceUrl = () => {
-  return process.env.NEXT_PUBLIC_API_URL;
+  if (typeof window !== "undefined") {
+    return "";
+  }
+  return SERVICE_SERVER_URL;
 };
